entity_recognition/Entity_recognition_json_utry_v2/requirements.txt
six
pandas
numpy
requests
foolnltk
flask
tornado
gunicorn
gevent

